Move Parser out of commands package
