+++ src/main/java/duke/Command.java
@@ -1,7 +1,5 @@
 package duke;
 
-import duke.commands.Parser;
-
 abstract public class Command {
     protected Parser parser = new Parser();
     protected String name;
