+++ src/main/java/duke/commands/Event.java
@@ -3,6 +3,7 @@ package duke.commands;
 import duke.Command;
 import duke.Duke;
 import duke.DukeException;
+import duke.Parser;
 
 import java.util.Map;
